fix(test): stop reaching a pure function through a component that needs a DOM
next-app (bun test) has been failing on branches whose diff touches no
front-end file at all -- twice in a row on one go-api-only PR, while main
passed at the identical next-app tree. The failure is always the same:

    ##[group] src/components/anime/continueWatchingTotal.test.ts
    # Unhandled error between tests
    ReferenceError: document is not defined
      at next/dist/shared/lib/deployment-id.js:33

continueWatchingTotal.test.ts tests resolveWatchingTotal, nine lines of
arithmetic over two numbers. It reached it by importing ContinueWatching,
which pulls FadeImage, which pulls next/image, whose deployment-id module
runs at import time:

    if (typeof window !== 'undefined') {
        deploymentId = document.documentElement.dataset.dplId;
    }

bun:test shares one process and one module cache across every test file,
and several files in this suite set globalThis.window while they run
(pendingSubscribe, SubscriptionSetProvider, api.buildHeaders). So whether
that import throws depends on which file happened to be executing when
some other file first pulled Next in -- ordering nothing in the suite
controls, which is why it presents as an intermittent failure on unrelated
diffs.

resolveWatchingTotal moves to continueWatchingState.ts, the sibling module
that already holds this component's Next-free logic, and the test imports
it from there. The component imports it back, so nothing about the
rendered output changes.

Verified as a before/after under the exact hazard rather than by watching
CI go green: with globalThis.window set and no document, the old import
shape fails 0 pass / 1 fail with that same ReferenceError, and the new one
passes 5 / 5. Full suite 1900 pass / 0 fail, tsc clean, build clean.

Note what this does NOT fix: any other test that reaches a component
module while another file has window set can hit the same wall. This
removes the one instance that was firing, and the function now carries a
comment saying why it lives where it does.
